+++ common/src/main/java/com/rimo/sfcr/mixin/WorldRendererMixin.java
@@ -1,13 +1,8 @@
 package com.rimo.sfcr.mixin;
 
-import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
-import com.rimo.sfcr.Client;
 import net.minecraft.client.render.WorldRenderer;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.client.world.ClientWorld;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.biome.Biome;
 import org.jetbrains.annotations.Nullable;
 import org.joml.Matrix4f;
 import org.spongepowered.asm.mixin.Mixin;
@@ -32,26 +27,6 @@ public abstract class WorldRendererMixin {
 			RENDERER.render(matrices, projectionMatrix, tickDelta, cameraX, cameraY, cameraZ, world);
 			ci.cancel();
 		}
-	}
-
-	/* - - NO CLOUD NO RAIN - - */
-
-	@WrapOperation(method = "renderWeather", at = @At(
-			value = "INVOKE",
-			target = "Lnet/minecraft/world/biome/Biome;getPrecipitation(Lnet/minecraft/util/math/BlockPos;)Lnet/minecraft/world/biome/Biome$Precipitation;"))
-	private Biome.Precipitation sfcr$redirectGetPrecipitationRain(Biome instance, BlockPos pos, Operation<Biome.Precipitation> original) {
-		if (Client.isNoCloudCovered(pos.getX(), pos.getY(), pos.getZ()))
-			return Biome.Precipitation.NONE;
-		return original.call(instance, pos);
-	}
-
-	@WrapOperation(method = "tickRainSplashing", at = @At(
-			value = "INVOKE",
-			target = "Lnet/minecraft/world/biome/Biome;getPrecipitation(Lnet/minecraft/util/math/BlockPos;)Lnet/minecraft/world/biome/Biome$Precipitation;"))
-	private Biome.Precipitation sfcr$redirectGetPrecipitationSplash(Biome instance, BlockPos pos, Operation<Biome.Precipitation> original) {
-		if (Client.isNoCloudCovered(pos.getX(), pos.getY(), pos.getZ()))
-			return Biome.Precipitation.NONE;
-		return original.call(instance, pos);
 	}
 
 }
