+++ common/src/main/java/com/rimo/sfcr/mixin/extra/WorldRendererMixin.java
@@ -1,0 +1,35 @@
+package com.rimo.sfcr.mixin.extra;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.rimo.sfcr.Client;
+import net.minecraft.client.render.WorldRenderer;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.biome.Biome;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(WorldRenderer.class)
+public abstract class WorldRendererMixin {
+
+	/* - - NO CLOUD NO RAIN - - */
+
+	@WrapOperation(method = "renderWeather", at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/biome/Biome;getPrecipitation(Lnet/minecraft/util/math/BlockPos;)Lnet/minecraft/world/biome/Biome$Precipitation;"))
+	private Biome.Precipitation sfcr$redirectGetPrecipitationRain(Biome instance, BlockPos pos, Operation<Biome.Precipitation> original) {
+		if (Client.isNoCloudCovered(pos.getX(), pos.getY(), pos.getZ()))
+			return Biome.Precipitation.NONE;
+		return original.call(instance, pos);
+	}
+
+	@WrapOperation(method = "tickRainSplashing", at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/biome/Biome;getPrecipitation(Lnet/minecraft/util/math/BlockPos;)Lnet/minecraft/world/biome/Biome$Precipitation;"))
+	private Biome.Precipitation sfcr$redirectGetPrecipitationSplash(Biome instance, BlockPos pos, Operation<Biome.Precipitation> original) {
+		if (Client.isNoCloudCovered(pos.getX(), pos.getY(), pos.getZ()))
+			return Biome.Precipitation.NONE;
+		return original.call(instance, pos);
+	}
+
+}
