+++ gradle.properties
@@ -4,7 +4,7 @@ minecraft_version=1.20.1
 enabled_platforms=fabric,forge
 
 archives_base_name=sfcr
-mod_version=1.8.2-mc1.20.1
+mod_version=1.9.0-mc1.20.1
 maven_group=com.rimo
 
 architectury_version=9.2.14
@@ -13,7 +13,6 @@ distant_horizon_api_version=5.1.0
 distant_horizon_version=2.4.5-b-1.20.1
 particle_rain_version=v4-beta.6+1.20.1-fabric
 serene_seasons_version=c1BdjabH
-glitchcore_version=25HLOiOl
 fabric_seasons_version=2.4.2-BETA+1.20
 
 fabric_loader_version=0.17.1
